backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
motor>=3.3.2
pymongo>=4.6.1
cryptography>=42.0.0
argon2-cffi>=23.1.0
PyJWT>=2.10.1
slowapi>=0.1.9
pydantic>=2.7.0
redis>=5.0.1
python-dotenv>=1.0.0
